Moved Time type to separate file so that types could access extremes.
